requirements.txt
pymysql
psycopg2
pandas
python-dotenv
seaborn